+++ jax/_src/tpu/linalg/svd.py
@@ -277,9 +277,17 @@ def _svd_tpu(a, *, full_matrices, compute_uv, subset_by_index, algorithm=None):
 def _svd_tpu_lowering_rule(
     ctx, operand, *, full_matrices, compute_uv, subset_by_index, algorithm=None
 ):
-  del algorithm  # unused
   operand_aval, = ctx.avals_in
   m, n = operand_aval.shape[-2:]
+
+  if algorithm is not None and algorithm not in [
+      lax_linalg.SvdAlgorithm.DEFAULT,
+      lax_linalg.SvdAlgorithm.POLAR,
+  ]:
+    raise NotImplementedError(
+        'Only the POLAR (which is also DEFAULT on TPU) SVD algorithm is'
+        ' supported on TPU.'
+    )
 
   if m == 0 or n == 0:
     return mlir.lower_fun(lax_linalg._empty_svd, multiple_results=True)(